docs(core): new year tips
